pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nameai"
version = "0.2.0"
description = "AI model that generates creative business/brand names from descriptions"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "sentencepiece>=0.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "wandb>=0.16",
    "tqdm>=4.66",
    "datasets>=2.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
nameai = "nameai.cli.app:main"

[tool.ruff]
line-length = 100
target-version = "py310"
